# Harden the semantics of the bounded scan queries