@enobayram 2023-01-04
I've also taken a closer look at the queries and their plans we were running before this PR as well as the ones we're running after:

## Baseline

Here's a baseline query that scans 1000000 rows of the `events` table while performing a substring search on the `qualname` column. Note that we're searching for the `transferz` string to make sure that it doesn't find anything and we just observe the scan performance:
``` SQL
=> EXPLAIN ANALYZE SELECT * FROM events WHERE height >= 3305260 AND qualname ILIKE '%transferz%';
                                                                 QUERY PLAN                                                                  
---------------------------------------------------------------------------------------------------------------------------------------------
 Index Scan using events_height_chainid_idx on events  (cost=0.56..629145.65 rows=1 width=381) (actual time=969.094..969.095 rows=0 loops=1)
   Index Cond: (height >= 3305260)
   Filter: ((qualname)::text ~~* '%transferz%'::text)
   Rows Removed by Filter: 1000507
 Planning Time: 0.319 ms
 JIT:
   Functions: 4
   Options: Inlining true, Optimization true, Expressions true, Deforming true
   Timing: Generation 0.601 ms, Inlining 5.718 ms, Optimization 29.791 ms, Emission 17.297 ms, Total 53.406 ms
 Execution Time: 969.759 ms
(10 rows)

=> SET jit = OFF;
SET
=> EXPLAIN ANALYZE SELECT * FROM events WHERE height >= 3305260 AND qualname ILIKE '%transferz%';
                                                                 QUERY PLAN                                                                  
---------------------------------------------------------------------------------------------------------------------------------------------
 Index Scan using events_height_chainid_idx on events  (cost=0.56..629192.06 rows=1 width=381) (actual time=959.071..959.072 rows=0 loops=1)
   Index Cond: (height >= 3305260)
   Filter: ((qualname)::text ~~* '%transferz%'::text)
   Rows Removed by Filter: 1000638
 Planning Time: 0.515 ms
 Execution Time: 959.103 ms
(6 rows)
```

So, with a simple query like this, we're able to scan roughly 1000 events per millisecond.

## Scanning with the `height`, `requestkey`, `idx` order

The query above cheats a little by only ordering by the `height` column, but in order to support the "cursor" behavior of continuing from where the previous query left off, we need to order by `height`, `requestkey`, `idx`. So the following query is probably a fairer baseline:
``` SQL
=> EXPLAIN ANALYZE SELECT count(*) FILTER (WHERE qualname ILIKE '%transferz%') FROM (SELECT * FROM events ORDER BY height DESC, requestkey, idx LIMIT 1000000) t;
                                                                                 QUERY PLAN                                                                                  
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Aggregate  (cost=3321866.83..3321866.84 rows=1 width=8) (actual time=1880.821..1880.822 rows=1 loops=1)
   ->  Limit  (cost=1119.87..3306866.83 rows=1000000 width=247) (actual time=0.192..1291.142 rows=1000000 loops=1)
         ->  Incremental Sort  (cost=1119.87..170671523.13 rows=51628393 width=247) (actual time=0.191..1242.253 rows=1000000 loops=1)
               Sort Key: events.height DESC, events.requestkey, events.idx
               Presorted Key: events.height
               Full-sort Groups: 16928  Sort Method: quicksort  Average Memory: 33kB  Peak Memory: 33kB
               Pre-sorted Groups: 6671  Sort Method: quicksort  Average Memory: 92kB  Peak Memory: 115kB
               ->  Index Scan using events_height_chainid_idx on events  (cost=0.56..166475846.58 rows=51628393 width=247) (actual time=0.018..498.194 rows=1000395 loops=1)
 Planning Time: 0.113 ms
 Execution Time: 1880.857 ms
(10 rows)
```
This time, we're scanning through ~500 rows per millisecond.

## Bounded scan offset query before this PR

``` SQL
=> EXPLAIN ANALYZE SELECT *                                                                      
FROM (
  SELECT *
       , ROW_NUMBER() OVER (ORDER BY height DESC, requestkey, idx) as scan_num
       , count(*) FILTER (WHERE qualname ILIKE '%transferz%') OVER
           ( ORDER BY height DESC, requestkey, idx 
             ROWS BETWEEN unbounded PRECEDING AND CURRENT ROW
           ) AS foundCnt
  FROM events
) as scanned_events
WHERE foundCnt = 3 OR scan_num = 1000000
LIMIT 1                                             
;
                                                                                       QUERY PLAN                                                                                        
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=1119.89..1457.06 rows=1 width=397) (actual time=2912.215..2912.217 rows=1 loops=1)
   ->  Subquery Scan on scanned_events  (cost=1119.89..173642856.17 rows=515000 width=397) (actual time=2912.214..2912.214 rows=1 loops=1)
         Filter: ((scanned_events.foundcnt = 3) OR (scanned_events.scan_num = 1000000))
         Rows Removed by Filter: 999999
         ->  WindowAgg  (cost=1119.89..172868420.21 rows=51629064 width=397) (actual time=0.133..2857.800 rows=1000000 loops=1)
               ->  WindowAgg  (cost=1119.89..171706766.27 rows=51629064 width=389) (actual time=0.118..1646.881 rows=1000001 loops=1)
                     ->  Incremental Sort  (cost=1119.89..170674184.99 rows=51629064 width=381) (actual time=0.111..1195.146 rows=1000002 loops=1)
                           Sort Key: events.height DESC, events.requestkey, events.idx
                           Presorted Key: events.height
                           Full-sort Groups: 16938  Sort Method: quicksort  Average Memory: 65kB  Peak Memory: 66kB
                           Pre-sorted Groups: 6668  Sort Method: quicksort  Average Memory: 500kB  Peak Memory: 607kB
                           ->  Index Scan using events_height_chainid_idx on events  (cost=0.56..166478446.68 rows=51629064 width=381) (actual time=0.023..523.716 rows=1000053 loops=1)
 Planning Time: 0.183 ms
 Execution Time: 2912.265 ms
(14 rows)
```

So, seems like the whole bounded scan machinery used to cost us around 35% scanning speed in the offset query and we could scan ~350 rows per millisecond.

## Bounded scan offset query after this PR
``` SQL
=> EXPLAIN ANALYZE SELECT *                                                                      
FROM (
  SELECT *
       , ROW_NUMBER() OVER (ORDER BY height DESC, requestkey, idx) as scan_num
       , count(*) FILTER (WHERE qualname ILIKE '%transferz%') OVER
           ( ORDER BY height DESC, requestkey, idx 
             ROWS BETWEEN unbounded PRECEDING AND CURRENT ROW
           ) AS foundCnt
  FROM events
  ORDER BY height DESC, requestkey, idx
  LIMIT 1000000
) as scanned_events
WHERE foundCnt = 3 OR scan_num = 1000000
ORDER BY height DESC, requestkey, idx
LIMIT 1                                             
;
                                                                                          QUERY PLAN                                                                                           
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=1119.89..1457.06 rows=1 width=397) (actual time=2765.847..2765.849 rows=1 loops=1)
   ->  Subquery Scan on scanned_events  (cost=1119.89..3364375.44 rows=9975 width=397) (actual time=2765.846..2765.847 rows=1 loops=1)
         Filter: ((scanned_events.foundcnt = 3) OR (scanned_events.scan_num = 1000000))
         Rows Removed by Filter: 999999
         ->  Limit  (cost=1119.89..3349375.44 rows=1000000 width=397) (actual time=0.125..2709.176 rows=1000000 loops=1)
               ->  WindowAgg  (cost=1119.89..172868420.21 rows=51629064 width=397) (actual time=0.125..2658.246 rows=1000000 loops=1)
                     ->  WindowAgg  (cost=1119.89..171706766.27 rows=51629064 width=389) (actual time=0.113..1525.837 rows=1000001 loops=1)
                           ->  Incremental Sort  (cost=1119.89..170674184.99 rows=51629064 width=381) (actual time=0.107..1110.402 rows=1000002 loops=1)
                                 Sort Key: events.height DESC, events.requestkey, events.idx
                                 Presorted Key: events.height
                                 Full-sort Groups: 16938  Sort Method: quicksort  Average Memory: 65kB  Peak Memory: 66kB
                                 Pre-sorted Groups: 6668  Sort Method: quicksort  Average Memory: 500kB  Peak Memory: 607kB
                                 ->  Index Scan using events_height_chainid_idx on events  (cost=0.56..166478446.68 rows=51629064 width=381) (actual time=0.042..487.930 rows=1000056 loops=1)
 Planning Time: 0.134 ms
 Execution Time: 2765.887 ms
(15 rows)
```

So, roughly the same performance.

## Bounded scan limit query before this PR
``` SQL
=> EXPLAIN ANALYZE SELECT *
FROM (
  SELECT ((qualname) ilike ('%transferz%')) AND (true) AS is_match, events.*
      , ROW_NUMBER() 
          OVER (ORDER BY height DESC, requestkey DESC, idx ASC) 
        AS scan_num 
  FROM events
  WHERE ((3306471, 'dCRwmpjFkj4qGJmAMHT_qsWFJJ1d6bW-Q4RJtnwhRqw', idx)) 
        > ((height, requestkey, 1)) 
  LIMIT 1000000
) AS t  
WHERE (((t.scan_num) = 1000000) OR (t.is_match)) 
LIMIT 20
;
                                                                                       QUERY PLAN                                                                                        
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=1119.98..1253.35 rows=20 width=390) (actual time=2188.369..2188.371 rows=1 loops=1)
   ->  Subquery Scan on t  (cost=1119.98..3351951.14 rows=502500 width=390) (actual time=2188.367..2188.369 rows=1 loops=1)
         Filter: ((t.scan_num = 1000000) OR t.is_match)
         Rows Removed by Filter: 999999
         ->  Limit  (cost=1119.98..3339451.14 rows=1000000 width=390) (actual time=0.531..2141.358 rows=1000000 loops=1)
               ->  WindowAgg  (cost=1119.98..171844867.45 rows=51475944 width=390) (actual time=0.530..2092.570 rows=1000000 loops=1)
                     ->  Incremental Sort  (cost=1119.98..170686658.71 rows=51475944 width=381) (actual time=0.502..1140.253 rows=1000006 loops=1)
                           Sort Key: events.height DESC, events.requestkey DESC, events.idx
                           Presorted Key: events.height
                           Full-sort Groups: 17106  Sort Method: quicksort  Average Memory: 66kB  Peak Memory: 66kB
                           Pre-sorted Groups: 6338  Sort Method: quicksort  Average Memory: 614kB  Peak Memory: 697kB
                           ->  Index Scan using events_height_chainid_idx on events  (cost=0.56..166505009.34 rows=51475944 width=381) (actual time=0.098..516.332 rows=1000029 loops=1)
                                 Index Cond: (height <= 3306471)
                                 Filter: (ROW(3306471, 'dCRwmpjFkj4qGJmAMHT_qsWFJJ1d6bW-Q4RJtnwhRqw'::text, idx) > ROW(height, (requestkey)::text, 1))
                                 Rows Removed by Filter: 72
 Planning Time: 0.452 ms
 Execution Time: 2188.447 ms
(17 rows)
```
The good news is that the limit query along with all the bounded scan machinery still scans ~500 rows per millisecond, so we're not causing any significant slowdown by the extra book keeping.

## Bounded scan limit query after this PR
``` SQL
=> EXPLAIN ANALYZE SELECT *                                                                                                                                                                                            
FROM (
  SELECT ((qualname) ilike ('%transferz%')) AND (true) AS is_match, events.*
      , ROW_NUMBER() 
          OVER (ORDER BY height DESC, requestkey DESC, idx ASC) 
        AS scan_num 
  FROM events
  WHERE ((3306471, 'dCRwmpjFkj4qGJmAMHT_qsWFJJ1d6bW-Q4RJtnwhRqw', idx)) 
        > ((height, requestkey, 1)) 
  LIMIT 1000000
) AS t  
WHERE (((t.scan_num) = 1000000) OR (t.is_match)) 
ORDER BY height DESC, requestkey DESC, idx 
LIMIT 20
;
                                                                                       QUERY PLAN                                                                                        
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=1119.98..1253.35 rows=20 width=390) (actual time=2276.571..2276.574 rows=1 loops=1)
   ->  Subquery Scan on t  (cost=1119.98..3351950.04 rows=502500 width=390) (actual time=2276.568..2276.570 rows=1 loops=1)
         Filter: ((t.scan_num = 1000000) OR t.is_match)
         Rows Removed by Filter: 999999
         ->  Limit  (cost=1119.98..3339450.04 rows=1000000 width=390) (actual time=0.389..2227.379 rows=1000000 loops=1)
               ->  WindowAgg  (cost=1119.98..171844500.50 rows=51475851 width=390) (actual time=0.388..2177.172 rows=1000000 loops=1)
                     ->  Incremental Sort  (cost=1119.98..170686293.86 rows=51475851 width=381) (actual time=0.362..1182.322 rows=1000006 loops=1)
                           Sort Key: events.height DESC, events.requestkey DESC, events.idx
                           Presorted Key: events.height
                           Full-sort Groups: 17106  Sort Method: quicksort  Average Memory: 66kB  Peak Memory: 66kB
                           Pre-sorted Groups: 6338  Sort Method: quicksort  Average Memory: 614kB  Peak Memory: 697kB
                           ->  Index Scan using events_height_chainid_idx on events  (cost=0.56..166504653.03 rows=51475851 width=381) (actual time=0.071..531.409 rows=1000029 loops=1)
                                 Index Cond: (height <= 3306471)
                                 Filter: (ROW(3306471, 'dCRwmpjFkj4qGJmAMHT_qsWFJJ1d6bW-Q4RJtnwhRqw'::text, idx) > ROW(height, (requestkey)::text, 1))
                                 Rows Removed by Filter: 72
 Planning Time: 0.523 ms
 Execution Time: 2276.662 ms
(17 rows)
```

So, roughly the same performance.